+++ domain/src/main/kotlin/no/nav/su/se/bakover/domain/journalpost/JournalpostClient.kt
@@ -3,9 +3,6 @@ package no.nav.su.se.bakover.domain.journalpost
 import arrow.core.Either
 import no.nav.su.se.bakover.domain.journal.JournalpostId
 
-/**
- * Saf = Sak og arkiv fasade
- */
 interface JournalpostClient {
     fun hentJournalpost(journalpostId: JournalpostId): Either<KunneIkkeHenteJournalpost, HentetJournalpost>
 }
